+++ test/rheo/doctest_test.exs
@@ -15,4 +15,5 @@ defmodule Rheo.DoctestTest do
   doctest Rheo.Backend.Ecto.Codec
   doctest Rheo.Telemetry
   doctest Rheo.Instance
+  doctest Rheo.Broadway
 end